requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.2
aiosmtplib==4.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
bitarray==3.1.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
coverage==7.6.12
cryptography==44.0.2
cytoolz==1.0.1
dependency-injector==4.46.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.11
flake8==7.1.2
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
influxdb==5.3.2
influxdb-client==1.48.0
iniconfig==2.0.0
isort==6.0.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
propcache==0.3.0
psutil==7.0.0
py-solc-x==2.0.3
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
reactivex==4.0.4
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.23.1
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
solc==0.0.0a0
solidity-parser==0.1.1
starlette==0.46.1
toolz==1.0.0
twilio==9.4.6
types-aiofiles==24.1.0.20241221
types-jsonschema==4.23.0.20241208
types-psutil==7.0.0.20250218
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
watchdog==6.0.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
